requirements.txt
opencv-python
numpy
requests
mediapipe
ultralytics
Pillow
torch
torchvision
protobuf
streamlit